Re-implement IPsec status with MVC concept
