+++ src/opnsense/service/conf/actions.d/actions_ipsec.conf
@@ -1,6 +1,6 @@
 [list.status]
 command:/usr/local/opnsense/scripts/ipsec/list_status.py
-parameters:
+parameters:%s --filter=%s --limit=%s --offset=%s --sort_by=%s --connection=%s
 type:script_output
 message:IPsec list status
 
@@ -9,12 +9,6 @@ command:/usr/local/opnsense/scripts/ipsec/list_leases.py
 parameters:
 type:script_output
 message:IPsec list ip address pools
-
-[list.database]
-command:/usr/local/opnsense/scripts/ipsec/list_dbs.py
-parameters:%s --filter=%s --limit=%s --offset=%s --sort_by=%s
-type:script_output
-message:IPsec list database
 
 [connect]
 command:/usr/local/opnsense/scripts/ipsec/connect.py
